Fixing the capitalization of yes?/no? when changing the default.
`default?` only returns whether or not the parameter was set, it does not return its value.

According to prompt.rb, `no?` already sets default to false, so it seems unnecessary to also check the type when creating the default labels.

Also changed the spec to respect that a changed `default` changes the capitalisation of both question and answer.

In the end, yes?{default:false} is the same as no? and no?{default:true} is the same as yes?.

Fixes #47.