Fixes error when backspace is 1st keypress on filtered select
Also happens when backspace is the 2nd keypress after several non-alpha keys.

Because frozen string literals are enabled, the empty string assigned to `@filter` on List instantiation is a frozen string. Most of the time, `List#keypress` unfreezes `@filter` because it overwrites it.
However, when backspace is the first keypress, `@filter.slice!` blows up because it's trying to modify a frozen string.

2 ways to fix this issue:
1. change List#keybackspace to overwrite `@filter` like the other methods that interact with `@filter`
2. instantiate `@filter` as an unfrozen string literal

I chose option 2 because it solves this specific problem and also solves any similar problems in any other methods (present or future) that interact with `@filter`.