add the ability to specify keys to select list items
This change updates List (Prompt#select) to add the ability to specify keypress
triggers for choices. It makes List respect the already-existing Choice#key, as
well as adding displaying of the keys with :enum, overriding of those displayed
keys with Choice#key_name, and modification of keypress behavior for List. The
default behavior of keypress == move to item is kept, but the ability to have a
keypress immediately select an item was added. This changeset also includes some
whitespace, comment, and spec fixups.